Reverse direction of dependencies between Problem(s) and Specification(s)
